syncval: Move ResourceAccessState to own file
The ResourceAccessState and related object define the "model" for Hazard
detection and state update for a given access range.  The range
abstraction is defined in the AccessContext.
