+++ BUILD.gn
@@ -237,6 +237,9 @@ vvl_sources = [
   "layers/stateless/sl_utils.cpp",
   "layers/stateless/sl_wsi.cpp",
   "layers/stateless/stateless_validation.h",
+  "layers/sync/sync_common.h",
+  "layers/sync/sync_model.cpp",
+  "layers/sync/sync_model.h",
   "layers/sync/sync_utils.cpp",
   "layers/sync/sync_utils.h",
   "layers/sync/sync_validation.cpp",
